+++ crates/ruma-common/CHANGELOG.md
@@ -18,6 +18,9 @@ Breaking changes:
 - `JsonType` was renamed to `CanonicalJsonType` to reflect that it only
   represents the possible types of a `CanonicalJsonValue`. It can also be
   accessed with `CanonicalJsonValue::json_type()`.
+- Refactor and improve the variants of `RedactionError`:
+  - `NotOfType` was renamed to `InvalidType` and provides more details about the
+    invalid field.
 
 Improvements:
 
@@ -38,6 +41,7 @@ Improvements:
   reallocate when possible.
 - Add `EventId::new_v2_or_v3()` to construct event IDs formats which are based
   on the event reference hash.
+- `CanonicalJsonValue::json_type()` allows to get the `JsonType` of a value.
 
 # 0.17.1